requirements.txt
requests>=2.31
matplotlib>=3.8
slack-sdk>=3.27
pyyaml>=6.0
